+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### [unreleased]
 * (files) the methods to interact with the database are now protected to make the class easier to extend
 * (hibernate) minor refactoring
+* (internal) make the in-memory RefreshTokenMap accessible to subclasses
 
 ### [4.4.2] (2021-05-06)
 * (all) Dependencies upgraded to latest